chore: split the util folder into smaller, related chunks and moved functionality out of index file
