+++ src/index.ts
@@ -1,48 +1,13 @@
-import { addProperties, indentLeft } from './utils';
-import { HEADER } from './constants';
-import {
-  Bookmark,
-  isBookmark,
-  Folder,
-  isFolder,
-  FolderProperties,
-} from './types';
+import { createBookmark } from './formatters/bookmark';
+import { createFolder } from './formatters/folder';
+import { indentLeft } from './utils';
+import { HEADER, AMOUNT_TO_PAD_BY } from './constants';
+import { Bookmark, isBookmark, Folder, isFolder } from './types';
 
-// TODO: configurable with an options object
-const AMOUNT_TO_PAD_BY = 4;
-
-const buildFolder = (folder: Folder) => {
-  const { add_date, last_modified } = folder;
-
-  const properties: FolderProperties = {
-    ...(add_date && { ADD_DATE: add_date }),
-    ...(last_modified && { LAST_MODIFIED: last_modified }),
-  };
-
-  return `<H3` + `${addProperties(properties)}` + `>${folder.name}</H3>`;
-};
-
-const createFolder = (folder: Folder, pad: string) =>
-  [
-    `${pad}<DL>`,
-    `${pad}  <P>`,
-    `${pad}    <DT>`,
-    `${pad}      ${buildFolder(folder)}`,
-    `${pad}    </DT>`,
-    // recursively call the build function
-    `${buildHtml(folder.children, pad.length + AMOUNT_TO_PAD_BY)}`,
-    `${pad}  </P>`,
-    `${pad}</DL>`,
-  ].join('\n');
-
-const createBookmark = (bookmark: Bookmark, pad: string) =>
-  [
-    `${pad}<DT>`,
-    `${pad}  <A HREF="${bookmark.href}">${bookmark.name}</A>`,
-    `${pad}</DT>`,
-  ].join('\n');
-
-const buildHtml = (content: (Bookmark | Folder)[], indent: number) =>
+export const buildHtml = (
+  content: (Bookmark | Folder)[],
+  indent: number,
+): string =>
   content
     .reduce((html: string[], current) => {
       if (isFolder(current)) {
